добавил стрельбу

+++ src/Renderer/VertexBufferLayout.cpp
@@ -4,7 +4,7 @@ namespace RenderEngine
 {
 	void VertexBufferLayout::addElementLayout(const int count, const bool normalize)
 	{
-		m_layoutElements.push_back({ count, GL_FLOAT, normalize, count * sizeof(GLfloat) });
+		m_layoutElements.push_back({ count, GL_FLOAT, normalize, count * static_cast<unsigned int>(sizeof(GLfloat)) });
 		m_stride += m_layoutElements.back().size;
 	}
